Move UpdateWidgetColorsAfterValidation from Utils to GtkUtils
